--- linked list/doc.go
@@ -1,6 +1,0 @@
-// linked list project doc.go
-
-/*
-linked list document
-*/
-package main
