+++ linked list/go.mod
@@ -1,0 +1,3 @@
+module testMod
+
+go 1.21.4
